requirements.txt
pandas
numpy
spacy
python-dotenv
supabase
scikit-learn
torch
rich
sentence-transformers==2.2.2
cachetools
